reorganised backend middleware
