+++ backend/routes/admin.js
@@ -1,11 +1,10 @@
 const         express = require('express');
 const adminController = require('../controllers/admin');
 const          router = express.Router();
-const      isLoggedIn = require('../middleware/isLoggedin')
 
-router.post('/add-listing',                 isLoggedIn, adminController.postAddListing);
-router.get('/user-listings',                isLoggedIn, adminController.getUserListings);
-router.put('/edit-listing/:listingId',      isLoggedIn, adminController.putEditListing);
-router.delete('/delete-listing/:listingId', isLoggedIn, adminController.deleteListing);
+router.post('/add-listing',                 adminController.postAddListing);
+router.get('/user-listings',                adminController.getUserListings);
+router.put('/edit-listing/:listingId',      adminController.putEditListing);
+router.delete('/delete-listing/:listingId', adminController.deleteListing);
 
 module.exports = router;
