+++ backend/routes/auth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 
 const authController = require('../controllers/auth');
-const { validateSignup } = require('../validation/auth');
+const { validateSignup } = require('../middleware/validateSignup');
 
 const router = express.Router();
