+++ backend/routes/chat.js
@@ -1,10 +1,9 @@
 const        express = require('express');
 const chatController = require('../controllers/chat');
-const     isLoggedIn = require('../middleware/isLoggedin');
 const         router = express.Router();
 
-router.post('/chat',                     isLoggedIn, chatController.postChat);
-router.get('/chats/',                    isLoggedIn, chatController.getChats);
-router.get('/chat/:sellerId/:listingId', isLoggedIn, chatController.findChat);
+router.post('/chat',                     chatController.postChat);
+router.get('/chats/',                    chatController.getChats);
+router.get('/chat/:sellerId/:listingId', chatController.findChat);
 
 module.exports = router;
