+++ backend/routes/message.js
@@ -1,9 +1,8 @@
 const           express = require('express');
 const messageController = require('../controllers/message');
-const        isLoggedIn = require('../middleware/isLoggedin');
 const            router = express.Router();
 
-router.post('/message',         isLoggedIn, messageController.postMessage);
-router.get('/messages/:chatId', isLoggedIn, messageController.getMessages);
+router.post('/message',         messageController.postMessage);
+router.get('/messages/:chatId', messageController.getMessages);
 
 module.exports = router;
