+++ backend/util/toObjectId.js
@@ -1,0 +1,3 @@
+const mongoose = require('mongoose');
+
+exports.toObjectId = (id) => new mongoose.Types.ObjectId(String(id));
